Made Some Pooling Improvements
I added a Debug Menu for Enemy Spawning to see the stats live. I also adjusted the enemy spawning max so it would now work. Also I made it so enemies truly only spawn when not in the view of the camera. Currently the game breaks over the 20 enemy limit which will need to be fixed