Enemy Animations
The enemies now have proper animations that switched depending what side the player is on. I also added a few things back to the scene that seemed to have been lost. In addition, I changed the enemies to now use a instance instead of finding a reference for the game manager